requirements.txt
pandas >= 0.25.0
matplotlib
pyspark
scipy
google-cloud-bigquery
